+++ BUILD
@@ -1,7 +1,3 @@
-exports_files([
-    "config.yaml",
-])
-
 java_library(
     name = "simulation-lib",
     srcs = glob(["*.java"]),
@@ -13,10 +9,11 @@ java_library(
     deps = [
         "//agents",
         "//common",
+        "//config",
+        "//yaml_tool",
         "@maven//:ch_qos_logback_logback_classic",
         "@maven//:commons_cli_commons_cli",
         "@maven//:org_slf4j_slf4j_api",
-        "//yaml_tool",
         "@graknlabs_client_java//:client-java",
         "@graknlabs_grabl_tracing//client",
         "@graknlabs_graql//java:graql",
@@ -32,7 +29,7 @@ java_binary(
     data = [
         "//data",
         "//schema",
-        "//:config.yaml",
+        "//config:config.yaml",
     ],
     main_class = "grakn.simulation.Simulation",
     runtime_deps = [":simulation-lib"],
