--- yaml_tool/AgentConfig.java
@@ -1,32 +1,0 @@
-package grakn.simulation.yaml_tool;
-
-public class AgentConfig {
-    private String name;
-    private Boolean run;
-    private Boolean sample;
-
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Boolean getSample() {
-        return sample;
-    }
-
-    public void setSample(Boolean sample) {
-        this.sample = sample;
-    }
-
-    public Boolean getRun() {
-        return run;
-    }
-
-    public void setRun(Boolean run) {
-        this.run = run;
-    }
-}
